move getDbPassword to own module
