+++ apps/auth-server/db.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { Sequelize } = require('sequelize');
+const getDbPassword = require('./utils/getDbPassword');
 
 let ssl = {};
 
@@ -14,16 +15,6 @@ if (process.env.DB_REQUIRE_SSL) {
 
 const dialectOptions = {
   ssl
-}
-
-const getDbPassword = async () => {
-  switch(process.env.DB_AUTH_METHOD) {
-    case 'azure-auth-token':
-      const { getAzureAuthToken } = require('./utils/azure')
-      return await getAzureAuthToken()
-    default:
-      return process.env.DB_PASSWORD
-  }
 }
 
 let sequelize = new Sequelize({
